Rudimentary type system for the MMTk harness scripting language.  Some new scripts that exploit the new features, plus some follow-up changes for issues revealed by the new scripts.

+++ MMTk/harness/src/org/mmtk/harness/lang/Intrinsics.java
@@ -68,7 +68,7 @@ public class Intrinsics {
    * @return
    */
   public static int random(Env env, int low, int high) {
-    return (int)(env.random().nextInt(high-low+1) + low);
+    return env.random().nextInt(high-low+1) + low;
   }
 
   /**
